My first commit

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -11,7 +11,7 @@
 int main(void)
 {
 	int first = 0, second = 1, next;
-	int f;
+	float f;
 
 	while (1)
 	{
